Changed position of side tentacles, fixed dodge armature bug, partially fixed animation override in the boss' flip coroutine,

+++ Project/Assets/Scripts/ScreenShake.cs
@@ -12,7 +12,7 @@ public class ScreenShake : MonoBehaviour
     public float decreaseVal = 5; //how long in seconds the shake lasts
     float shakeAmount;
 
-    public void shakeCamera( int magnitude )
+    public void shakeCamera( float magnitude )
     {
         Debug.Log("CameraShake " + magnitude);
         shakeAmount = magnitude * shakeMultiplier;
